+++ Cesium3DTiles/include/Cesium3DTiles/Tileset.h
@@ -99,7 +99,7 @@ namespace Cesium3DTiles {
          * loaded bytes is greater than this value, tiles will be unloaded until the total is under
          * this number or until only required tiles remain, whichever comes first.
          */
-        size_t maximumCachedBytes = 512 * 1024 * 1024;
+        int64_t maximumCachedBytes = 512 * 1024 * 1024;
 
         /**
          * @brief A table that maps the camera height above the ellipsoid to a fog density. Tiles that are in full fog are culled.
@@ -299,7 +299,7 @@ namespace Cesium3DTiles {
         /**
          * @brief Gets the total number of bytes of tile and raster overlay data that are currently loaded.
          */
-        size_t getTotalDataBytes() const noexcept;
+        int64_t getTotalDataBytes() const noexcept;
 
     private:
         struct TraversalDetails {
@@ -500,7 +500,7 @@ namespace Cesium3DTiles {
 
         RasterOverlayCollection _overlays;
 
-        size_t _tileDataBytes;
+        int64_t _tileDataBytes;
 
         static void addTileToLoadQueue(std::vector<LoadRecord>& loadQueue, const FrameState& frameState, Tile& tile, double distance);
         static void processQueue(std::vector<Tileset::LoadRecord>& queue, std::atomic<uint32_t>& loadsInProgress, uint32_t maximumLoadsInProgress);
